#36 [fix] [rename] 일정 페이지의 파일 이름 변경, 라우터 설정, 네비게이션 설정
