+++ package.json
@@ -30,9 +30,8 @@
     "core-js": "3.33.0",
     "date-fns": "^3.6.0",
     "handsontable": "^14.3.0",
-    // "vue": "^3.4.27",
+    "vue": "^3.4.27",
     "sockjs-client": "^1.6.1",
-    "vue": "3.3.4",
     "vue-router": "4.2.5",
     "vue-toastification": "^2.0.0-rc.5",
     "vue3-cookies": "^1.0.6",
