+++ src/router/index.js
@@ -1,7 +1,7 @@
 import { createRouter, createWebHistory } from "vue-router";
 import Dashboard from "../views/Dashboard.vue";
 import ProjectMember from "../views/ProjectMember.vue";
-import Billing from "../views/Billing.vue";
+import Schedules from "../views/Schedules.vue";
 import Notifications from "../views/Notifications.vue";
 import SignIn from "../views/SignIn.vue";
 import MaterialSchedule from "@/components/MaterialSchedule.vue";
@@ -30,11 +30,6 @@ const routes = [
     component: ProjectMember,
   },
   {
-    path: "/billing",
-    name: "일정",
-    component: Billing,
-  },
-  {
     path: "/notifications",
     name: "알림",
     component: Notifications,
@@ -53,6 +48,11 @@ const routes = [
     path: "/requirements/register",
     name: "RegisterRequirement",
     component: RegisterRequirement,
+  },
+  {
+    path: "/shcedules",
+    name: "일정",
+    component: Schedules,
   },
   {
     path: "/schedules/details/:scheduleId",
